Suppress a few signedness warnings

+++ assemble.c
@@ -378,7 +378,7 @@ int32_t assemble(int32_t segment, int32_t offset, int bits, uint32_t cp,
                 l = len;
                 while (l > 0) {
                     int32_t m =
-                        fread(buf, 1, (l > sizeof(buf) ? sizeof(buf) : l),
+                        fread(buf, 1, (l > (int32_t) sizeof(buf) ? (int32_t) sizeof(buf) : l),
                               fp);
                     if (!m) {
                         /*
